feat: silence all alerts for training except erroneous questionnaires
refs: BLAIS5-3953
