+++ lib/send_alerts.py
@@ -5,8 +5,8 @@ from typing import List
 from lib.alerter import Alerter
 from lib.cloud_functions import InvalidCloudFunctionEvent, parse_event
 from lib.cloud_logging import parse_log_entry
-from lib.filters.all_preprod_alerts_except_erroneous_questionnaire_filter import (
-    all_preprod_alerts_except_erroneous_questionnaire_filter,
+from lib.filters.all_preprod_and_training_alerts_except_erroneous_questionnaire_filter import (
+    all_preprod_and_training_alerts_except_erroneous_questionnaire_filter,
 )
 from lib.filters.sandbox_filter import sandbox_filter
 from lib.filters.auditlog_filter import auditlog_filter
@@ -29,7 +29,7 @@ from lib.log_processor import process_log_entry
 def log_entry_skipped(log_entry: ProcessedLogEntry):
     filters = [
         sandbox_filter,
-        all_preprod_alerts_except_erroneous_questionnaire_filter,
+        all_preprod_and_training_alerts_except_erroneous_questionnaire_filter,
         osconfig_agent_filter,
         auditlog_filter,
         agent_connect_filter,
